+++ changelog.md
@@ -1,6 +1,10 @@
 # 1.0.0-alpha.2
 - UI: Make missing lua.dll popup better
 - UI: Plugins that were never used show up in Plugins list
+- UI: Add plugin indicator
+- UI: Add delete script/plugin buttons
+- API: Add `native` key to `PluginMetadata`
+- API: Add `path` key to `PluginMetadata`
 
 # 1.0.0-alpha.1
 - Initial release